+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * fix selection + css variables (know bug in MS Edge)
 * fix apply/call issue that was causing Android 2.3 to crash
 * fix context menu on selected text (the selected text was cleared)
+* allow to overwrite keymap defined in terminal (not only the ones defined in cmd).
 
 
 ## 1.8.0